Conflicts:
	kexi/main/KexiMainWindow.cpp
	kexi/main/KexiMainWindow_p.h
	krita/image/kis_image.cc
	krita/image/kis_node_graph_listener.cpp
	krita/image/kis_paint_device.cc
	krita/image/kis_paint_layer.cc
	krita/image/kis_strokes_queue.cpp
	krita/image/kis_update_scheduler.cpp
	krita/image/tests/kis_paint_layer_test.cpp
	krita/image/tests/scheduler_utils.h
	krita/plugins/paintops/colorsmudge/kis_overlay_mode_option.h
	krita/plugins/paintops/libpaintop/kis_curve_option.cpp
	krita/ui/KisDocument.cpp
	krita/ui/KisMainWindow.cpp
	krita/ui/KisPart.cpp
	krita/ui/canvas/kis_canvas2.cpp
	krita/ui/kis_zoom_manager.cc

+++ krita/plugins/paintops/chalk/kis_chalk_paintop_settings.cpp
@@ -51,7 +51,7 @@ QPainterPath KisChalkPaintOpSettings::brushOutline(const KisPaintInformation &in
 
         QPainterPath tiltLine;
         QLineF tiltAngle(QPointF(0.0,0.0), QPointF(0.0,size));
-        tiltAngle.setLength(qMax(size*0.5, qreal(50.0)) * (1 - info.tiltElevation(info, 60.0, 60.0, true)));
+        tiltAngle.setLength(qMax(size*qreal(0.5), qreal(50.0)) * (1 - info.tiltElevation(info, 60.0, 60.0, true)));
         tiltAngle.setAngle((360.0 - fmod(KisPaintInformation::tiltDirection(info, true) * 360.0 + 270.0, 360.0))-3.0);
         tiltLine.moveTo(tiltAngle.p1());
         tiltLine.lineTo(tiltAngle.p2());
